refactor: allUser 파일명 수정
